Day 16 part 2 solution fix
Works for the regular input now. It's quite a bit slower and part 2 could probably be optimized based on the previous version of this solution by applying the same fix. For now, part 2 simulates all possible ways to split the valves into two sets which results in ~16k simulations which is quite manageable, especially with multithreading